requirements.txt
websockets==15.0.1
asyncio
pyaudio==0.2.11
numpy==1.24.3
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
aiohttp==3.9.1
openai==1.92.3